Struct maybe_utf8::MaybeUtf8Buf

pub struct MaybeUtf8Buf {
    // some fields omitted
}

Byte container optionally encoded as UTF-8. It might be either...

Definitely UTF-8-encoded string, or

Bytes. It may be encoded in UTF-8 or other encodings, or it may be simply invalid.

Methods

impl MaybeUtf8Buf

fn new() -> MaybeUtf8Buf

Creates a new empty MaybeUtf8Buf value (which is, naturally, encoded in UTF-8).

fn from_str(s: String) -> MaybeUtf8Buf

Creates a MaybeUtf8Buf value from an owned String.

fn from_bytes(v: Vec<u8>) -> MaybeUtf8Buf

Creates a MaybeUtf8Buf value from an owned Vec of u8 bytes.

fn as_bytes<'a>(&'a self) -> &'a [u8]

Returns a slice of underlying bytes. It might or might not be encoded in UTF-8.

fn as_str<'a>(&'a self) -> Option<&'a str>

Returns a string slice encoded in UTF-8 if possible. It returns None if the underlying bytes are not encoded in UTF-8.

fn map_as_cow<'a, F, T>(&'a self, to_cow: F) -> Cow<'a, str> where F: FnMut(&'a [u8]) -> T, T: IntoCow<'a, str>

Returns a Cow string which represents the current MaybeUtf8Slice. It may call given to_cow function to get a Cow string out of the bytes. to_cow function itself may return a String or &str compatible to Cow string.

fn as_cow_lossy<'a>(&'a self) -> Cow<'a, str>

Returns a Cow string which represents the current MaybeUtf8Slice. Any invalid UTF-8 sequences are replaced by U+FFFD, as like String::from_utf8_lossy.

fn to_slice<'a>(&'a self) -> MaybeUtf8Slice<'a>

Returns a MaybeUtf8Slice borrowed from this MaybeUtf8Buf.

fn into_str(self) -> Result<String, MaybeUtf8Buf>

Tries to convert a MaybeUtf8Buf into a String. If there is an invalid UTF-8 sequence it returns the original MaybeUtf8Buf back.

fn map_into_str<F>(self, into_str: F) -> String where F: FnMut(Vec<u8>) -> String

Converts a MaybeUtf8Buf into a String. It may call given into_str function to get a String out of the bytes.

fn into_str_lossy(self) -> String

Converts a MaybeUtf8Buf into a String. Any invalid UTF-8 sequences are replaced by U+FFFD, as like String::from_utf8_lossy.

fn into_bytes(self) -> Vec<u8>

Converts a MaybeUtf8Buf into a Vec of u8 bytes.

fn len(&self) -> usize

Returns a byte length of the MaybeUtf8Buf value.